api/python/requirements.txt
# Taste of Gratitude Python API - Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
squareup>=35.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
python-multipart>=0.0.6
